moved sector stuff into events
