+++ QSB/Events/PlayerSectorEvent.cs
@@ -1,4 +1,6 @@
 ﻿using QSB.Messaging;
+using QSB.TransformSync;
+using QSB.Utility;
 
 namespace QSB.Events
 {
@@ -8,7 +10,7 @@ namespace QSB.Events
 
         public override void SetupListener()
         {
-            GlobalMessenger<uint, int, string>.AddListener("QSBPlayerSectorChange", (netId, id, name) => SendEvent(
+            GlobalMessenger<uint, int, string>.AddListener("QSBSectorChange", (netId, id, name) => SendEvent(
                 new SectorMessage {
                     SenderId = netId,
                     SectorId = id,
@@ -18,7 +20,17 @@ namespace QSB.Events
 
         public override void OnReceive(SectorMessage message)
         {
-            throw new System.NotImplementedException();
+            var sector = SectorSync.LocalInstance.FindSectorByName((Sector.Name)message.SectorId, message.SectorName);
+
+            if (sector == null)
+            {
+                DebugLog.ToScreen($"Sector {message.SectorName},{(Sector.Name)message.SectorId} not found!");
+                return;
+            }
+
+            var transformSync = PlayerRegistry.GetTransformSync(message.SenderId);
+            DebugLog.ToScreen($"{transformSync.GetType().Name} of ID {message.SenderId} set to {message.SectorName}");
+            transformSync.ReferenceTransform = sector.transform;
         }
     }
 }
